added security feature to make sure you have adequate permissions to delete a group

+++ GPS With Friends/web/deleteGroup.jsp
@@ -16,25 +16,29 @@
      
     <%-- start web service invocation --%><hr/>
     <%
-    try {
-	edu.uoregon.cs.client.GPSwfriends_Service service = new edu.uoregon.cs.client.GPSwfriends_Service();
-	edu.uoregon.cs.client.GPSwfriends port = service.getGPSwfriendsPort();
-	 // TODO initialize WS operation arguments here
-	int gid = Integer.parseInt(session.getAttribute("gid").toString());
-	// TODO process result here
-	edu.uoregon.cs.client.Status result = port.deleteGroup(gid);
-        if (result.isSuccess()){
-            out.println("Group has been succesfully deleted");
+    if (request.getParameter("delete").equals("yes")) {
+        try {
+            edu.uoregon.cs.client.GPSwfriends_Service service = new edu.uoregon.cs.client.GPSwfriends_Service();
+            edu.uoregon.cs.client.GPSwfriends port = service.getGPSwfriendsPort();
+             // TODO initialize WS operation arguments here
+            int gid = Integer.parseInt(session.getAttribute("gid").toString());
+            // TODO process result here
+            edu.uoregon.cs.client.Status result = port.deleteGroup(gid);
+            if (result.isSuccess()){
+                out.println("Group has been succesfully deleted");
             %>
             <form name="back" action="main.jsp">
                     <input type="submit" value="Back"/>
             </form>
     <%
-        } else {
-            out.println(result.getError());
+            } else {
+                out.println(result.getError());
+            }
+        } catch (Exception ex) {
+            // TODO handle custom exceptions here
         }
-    } catch (Exception ex) {
-	// TODO handle custom exceptions here
+    } else {
+        out.println("You do not have permission to delete this group. If you think you do have permission please go to the group page and use the delete button their");
     }
     %>
     <%-- end web service invocation --%><hr/>
